+++ examples/common.rs
@@ -73,7 +73,7 @@ mod app {
         // This task is only spawned once in `init`, hence this task will run
         // only once
 
-        hprintln!("foo").ok();
+        hprintln!("foo");
     }
 
     // Software task, also not bound to a hardware interrupt
@@ -81,7 +81,7 @@ mod app {
     // The resources `s1` and `s2` are shared between all other tasks.
     #[task(shared = [s1, s2], local = [l2])]
     fn bar(_: bar::Context) {
-        hprintln!("bar").ok();
+        hprintln!("bar");
 
         // Run `bar` once per second
         bar::spawn_after(1.secs()).unwrap();
@@ -97,6 +97,6 @@ mod app {
         // Note that RTIC does NOT clear the interrupt flag, this is up to the
         // user
 
-        hprintln!("UART0 interrupt!").ok();
+        hprintln!("UART0 interrupt!");
     }
 }