+++ examples/generics.rs
@@ -32,19 +32,22 @@ mod app {
 
     #[task(binds = UART0, shared = [shared], local = [state: u32 = 0])]
     fn uart0(c: uart0::Context) {
-        hprintln!("UART0(STATE = {})", *c.local.state).unwrap();
+        hprintln!("UART0(STATE = {})", *c.local.state);
 
         // second argument has type `shared::shared`
         super::advance(c.local.state, c.shared.shared);
 
         rtic::pend(Interrupt::UART1);
 
+        // Exit moved after nop to ensure that rtic::pend gets
+        // to run before exiting
+        cortex_m::asm::nop();
         debug::exit(debug::EXIT_SUCCESS); // Exit QEMU simulator
     }
 
     #[task(binds = UART1, priority = 2, shared = [shared], local = [state: u32 = 0])]
     fn uart1(c: uart1::Context) {
-        hprintln!("UART1(STATE = {})", *c.local.state).unwrap();
+        hprintln!("UART1(STATE = {})", *c.local.state);
 
         // second argument has type `shared::shared`
         super::advance(c.local.state, c.shared.shared);
@@ -61,5 +64,5 @@ fn advance(state: &mut u32, mut shared: impl Mutex<T = u32>) {
         (old, *shared)
     });
 
-    hprintln!("shared: {} -> {}", old, new).unwrap();
+    hprintln!("shared: {} -> {}", old, new);
 }